Move test to /test/browser/spec/ol/format/wfs.test.js file
